pyproject.toml
[project]
name = "gateway"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.5",
    "flask>=3.1.2",
    "gunicorn>=23.0.0",
    "pymysql>=1.1.2",
    "python-dotenv>=1.1.1",
    "redis>=6.4.0",
    "rq>=2.6.0",
    "rq-scheduler>=0.14.0",
    "sqlalchemy>=2.0.43",
    "psutil>=6.1.0",
    "pynostr>=0.7.0",
    "grpcio>=1.60.0",
    "grpcio-tools>=1.60.0",
    "cryptography>=41.0.0",
    "protobuf>=4.25.0",
    "ecdsa>=0.19.0",
    "base58>=2.1.1",
    "bech32>=1.2.0",
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "prometheus-client>=0.23.1",
]
